Trying to incorporate more general assets like bump maps

+++ iviz/Assets/Core/Utils/SdfUtils.cs
@@ -4,12 +4,12 @@ namespace Iviz.Core
 {
     public static class SdfUtils
     {
-        public static Vector3 ToVector3(this Sdf.Vector3f v)
+        public static Vector3 ToVector3(this Sdf.Vector3d v)
         {
             return new Vector3((float)v.X, (float)v.Y, (float)v.Z).Ros2Unity();
         }
 
-        static Quaternion ToQuaternion(this Sdf.Vector3f v)
+        static Quaternion ToQuaternion(this Sdf.Vector3d v)
         {
             return new Vector3((float)v.X, (float)v.Y, (float)v.Z).RosRpy2Unity();
         }
@@ -28,7 +28,7 @@ namespace Iviz.Core
         {
             return new Color32(v.R, v.G, v.B, v.A);
         }
-        
+
         public static Vector3 ToVector3(this Msgs.IvizMsgs.Vector3f v)
         {
             return new Vector3(v.X, v.Y, v.Z).Ros2Unity();
